grade fixes

+++ gf/db/migrate/20151104133110_create_submissions.rb
@@ -5,7 +5,7 @@ class CreateSubmissions < ActiveRecord::Migration
       t.timestamps null: false
       t.string :title
       t.string :name
-      t.integer :grade
+      t.float :grade
       t.text :notecard
       t.string :vocab_english_1
       t.string :vocab_english_2
